Use native platform detection for DEB and RPM
Basing the package platform against the kernel machine name
has a lot of pitfalls, especially on DEB based systems. Instead
of having lots of code to try and figure out the correct
architecture, this just asks the packaing tools.
On DEB based systems the command is:
  dpkg --print-architecture
For RPM based systems the command is:
  rpmspec -E "%{_arch}"
Because omnibus already uses the native packaing tools, the
require binaries will already be available. This is really
only an issue on RPM based systems because with the advent of
dnf, rpmspec is no longer part of the rpm package but is instead
it's own package. That package is included when the rpm build
tools are installed. dpkg is always available on DEB based
systems.